--- a/plumbing/format/gitignore/dir_test.go
+++ b/plumbing/format/gitignore/dir_test.go
@@ -44,7 +44,9 @@ func (s *MatcherSuite) SetUpTest(c *C) {
 	c.Assert(err, IsNil)
 	_, err = f.Write([]byte("ignore.crlf\r\n"))
 	c.Assert(err, IsNil)
-	_, err = f.Write([]byte("ignore_dir\n"))
+	_, err = f.Write([]byte("/ignore_dir\n"))
+	c.Assert(err, IsNil)
+	_, err = f.Write([]byte("nested/ignore_dir\n"))
 	c.Assert(err, IsNil)
 	err = f.Close()
 	c.Assert(err, IsNil)
@@ -67,6 +69,17 @@ func (s *MatcherSuite) SetUpTest(c *C) {
 	_, err = fs.Create("ignore_dir/file")
 	c.Assert(err, IsNil)
 	err = f.Close()
+	c.Assert(err, IsNil)
+
+	err = fs.MkdirAll("nested/ignore_dir", os.ModePerm)
+	c.Assert(err, IsNil)
+	f, err = fs.Create("nested/ignore_dir/.gitignore")
+	c.Assert(err, IsNil)
+	_, err = f.Write([]byte("!file\n"))
+	c.Assert(err, IsNil)
+	err = f.Close()
+	c.Assert(err, IsNil)
+	_, err = fs.Create("nested/ignore_dir/file")
 	c.Assert(err, IsNil)
 
 	err = fs.MkdirAll("another", os.ModePerm)
@@ -280,13 +293,14 @@ func (s *MatcherSuite) SetUpTest(c *C) {
 
 func (s *MatcherSuite) TestDir_ReadPatterns(c *C) {
 	checkPatterns := func(ps []Pattern) {
-		c.Assert(ps, HasLen, 7)
+		c.Assert(ps, HasLen, 8)
 		m := NewMatcher(ps)
 
 		c.Assert(m.Match([]string{"exclude.crlf"}, true), Equals, true)
 		c.Assert(m.Match([]string{"ignore.crlf"}, true), Equals, true)
 		c.Assert(m.Match([]string{"vendor", "gopkg.in"}, true), Equals, true)
 		c.Assert(m.Match([]string{"ignore_dir", "file"}, false), Equals, true)
+		c.Assert(m.Match([]string{"nested", "ignore_dir", "file"}, false), Equals, true)
 		c.Assert(m.Match([]string{"vendor", "github.com"}, true), Equals, false)
 		c.Assert(m.Match([]string{"multiple", "sub", "ignores", "first", "ignore_dir"}, true), Equals, true)
 		c.Assert(m.Match([]string{"multiple", "sub", "ignores", "second", "ignore_dir"}, true), Equals, true)